LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/fill-and-stroke-styles/2d.gradient.radial.negative-expected.txt

2d.gradient.radial.negative

createRadialGradient() throws INDEX_SIZE_ERR if either radius is negative


PASS createRadialGradient() throws INDEX_SIZE_ERR if either radius is negative
